Added 3D model, added some very basic and terrible animations, updated script to be a bit more optimised and also work with animations.
I decided to adapt Eliotts code to fit a 3D model. It is still using some 2D collisions which is fine but the animation needed to be fitted into the code some how, so i reworked a big part of his code. Still kept some variables and logic from the previous code so that it is still somewhat the same.