Implemented Melee Attack, Improved Crouching behaivour.
I added a melee attack to the character, the player can do both melee and ranged attacks. This can be changed by setting the boolean "Using Ranged Weapon" to true for ranged attacks and to false for melee attacks. This will be changed in the future once we have set our minds on exacly melee. Additionally, i have added the behaivour of not being able to uncrouch yourself while colliding with something above you. This way, we can create small crawspaces where crouching would work consistently. I havent seen any oddities with the behaivour yet, but still, i dont think it works 100% of the time.